- renames combo_box_reset directory to use_combo_box_reset
