+++ x-pack/plugins/security_solution/public/common/components/use_combo_box_reset/index.tsx
